launch backend jobs from web GUI
