+++ src/Sir.CommonCrawl/Sir.CommonCrawl.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.6" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Sir.Search\Sir.Search.csproj" />
+    <ProjectReference Include="..\Sir\Sir.Core.csproj" />
+  </ItemGroup>
+
+</Project>
