+++ src/Sir.DbUtil/Program.cs
@@ -5,12 +5,11 @@ using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
-using System.Net;
 using System.Net.Http;
 using System.Text;
-using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
+using Sir.CommonCrawl;
 using Sir.Search;
 
 namespace Sir.DbUtil
@@ -59,24 +58,31 @@ namespace Sir.DbUtil
             {
                 // Ex: download_wat CC-MAIN-2019-51 d:\ cc_wat 0 1
 
-                DownloadAndIndexWat(args, model, loggerFactory, logger);
+                CCHelper.DownloadAndIndexWat(
+                    commonCrawlId:args[1],
+                    workingDirectory:args[2],
+                    collectionName:args[3],
+                    skip: int.Parse(args[4]),
+                    take: int.Parse(args[5]),
+                    model, 
+                    logger);
             }
             else if (command == "write_wet")
             {
                 // Ex: D:\CC-MAIN-2019-43\segments\1570986647517.11\wet\CC-MAIN-20191013195541-20191013222541-00000.warc.wet.gz
-                WriteWet(args, model, loggerFactory);
+                WriteWet(args, model, logger);
             }
             else if (command == "truncate")
             {
-                Truncate(args, model, loggerFactory);
+                Truncate(args, model, logger);
             }
             else if (command == "truncate-index")
             {
-                TruncateIndex(args, model, loggerFactory);
+                TruncateIndex(args, model, logger);
             }
             else if (command == "optimize")
             {
-                Optimize(args, model, loggerFactory);
+                Optimize(args, model, logger);
             }
             else
             {
@@ -86,14 +92,14 @@ namespace Sir.DbUtil
             logger.LogInformation($"executed {command}");
         }
 
-        private static void Optimize(string[] args, BocModel model, ILoggerFactory loggerFactory)
+        private static void Optimize(string[] args, BocModel model, ILogger logger)
         {
             var collection = args[1];
             var skip = int.Parse(args[2]);
             var take = int.Parse(args[3]);
             var batchSize = int.Parse(args[4]);
 
-            using (var sessionFactory = new SessionFactory(new KeyValueConfiguration("sir.ini"), model, loggerFactory))
+            using (var sessionFactory = new SessionFactory(new KeyValueConfiguration("sir.ini"), model, logger))
             {
                 sessionFactory.Optimize(
                     collection, 
@@ -105,7 +111,7 @@ namespace Sir.DbUtil
             }
         }
 
-        private static void WriteWet(string[] args, IStringModel model, ILoggerFactory logger)
+        private static void WriteWet(string[] args, IStringModel model, ILogger logger)
         {
             var fileName = args[1];
             var collectionId = "cc_wet".ToHash();
@@ -190,217 +196,6 @@ namespace Sir.DbUtil
             }
         }
 
-        private static void DownloadAndIndexWat(string[] args, IStringModel model, ILoggerFactory logger, ILogger log)
-        {
-            var ccName = args[1];
-            var workingDir = args[2];
-            var collection = args[3];
-            var skip = int.Parse(args[4]);
-            var take = int.Parse(args[5]);
-            var pathsFileName = $"{ccName}/wat.paths.gz";
-            var localPathsFileName = Path.Combine(workingDir, pathsFileName);
-
-            if (!File.Exists(localPathsFileName))
-            {
-                var url = $"https://commoncrawl.s3.amazonaws.com/crawl-data/{pathsFileName}";
-
-                log.LogInformation($"downloading {url}");
-
-                if (!Directory.Exists(Path.GetDirectoryName(localPathsFileName)))
-                {
-                    Directory.CreateDirectory(Path.GetDirectoryName(localPathsFileName));
-                }
-
-                using (var client = new WebClient())
-                {
-                    client.DownloadFile(url, localPathsFileName);
-                }
-
-                log.LogInformation($"downloaded {localPathsFileName}");
-            }
-
-            log.LogInformation($"processing {localPathsFileName}");
-
-            Task writeTask = null;
-            var took = 0;
-            var skipped = 0;
-
-            foreach (var watFileName in ReadAllLinesGromGz(localPathsFileName))
-            {
-                if (skip > skipped)
-                {
-                    skipped++;
-                    continue;
-                }
-
-                if (took++ == take)
-                {
-                    break;
-                }
-
-                var localWatFileName = Path.Combine(workingDir, watFileName);
-
-                if (!File.Exists(localWatFileName))
-                {
-                    var url = $"https://commoncrawl.s3.amazonaws.com/{watFileName}";
-
-                    log.LogInformation($"downloading {url}");
-
-                    if (!Directory.Exists(Path.GetDirectoryName(localWatFileName)))
-                    {
-                        Directory.CreateDirectory(Path.GetDirectoryName(localWatFileName));
-                    }
-
-                    using (var client = new WebClient())
-                    {
-                        client.DownloadFile(url, localWatFileName);
-                    }
-
-                    log.LogInformation($"downloaded {localWatFileName}");
-                }
-
-                var refFileName = watFileName.Replace(".wat", "").Replace("/wat", "/warc");
-
-                //log.LogInformation($"processing {localWatFileName}");
-                //WriteWatSegment(localWatFileName, collection, model, logger, log, refFileName);
-
-                if (writeTask != null)
-                {
-                    log.LogInformation($"synchronizing write");
-
-                    writeTask.Wait();
-                }
-
-                writeTask = Task.Run(() =>
-                {
-                    log.LogInformation($"processing {localWatFileName}");
-
-                    WriteWatSegment(localWatFileName, collection, model, logger, log, refFileName);
-                });
-            }
-        }
-
-        private static void WriteWatSegment(
-            string fileName, 
-            string collection, 
-            IStringModel model, 
-            ILoggerFactory log, 
-            ILogger logger,
-            string refFileName)
-        {
-            var documents = ReadWatFile(fileName, refFileName);
-            var collectionId = collection.ToHash();
-            var time = Stopwatch.StartNew();
-            var storedFieldNames = new HashSet<string>
-            {
-                "title","description", "url", "filename"
-            };
-            var indexedFieldNames = new HashSet<string>
-            {
-                "title","description", "url"
-            };
-
-            using (var sessionFactory = new SessionFactory(new KeyValueConfiguration("sir.ini"), model, log))
-            {
-                sessionFactory.Write(
-                            new WriteJob(
-                                collectionId, 
-                                documents, 
-                                model,
-                                storedFieldNames,
-                                indexedFieldNames),
-                            reportSize:1000);
-            }
-
-            logger.LogInformation($"indexed {fileName} in {time.Elapsed}");
-        }
-
-        private static IEnumerable<IDictionary<string, object>> ReadWatFile(string fileName, string refFileNae)
-        {
-            using (var fs = new FileStream(fileName, FileMode.Open, FileAccess.Read, FileShare.Read, 4096, FileOptions.SequentialScan))
-            using (var zip = new GZipStream(fs, CompressionMode.Decompress))
-            using (var reader = new StreamReader(zip, Encoding.UTF8))
-            {
-                var line = reader.ReadLine();
-
-                while (line != null)
-                {
-                    if (line.StartsWith('{'))
-                    {
-                        var doc = JsonConvert.DeserializeObject<Dictionary<string, object>>(
-                            line,
-                            new JsonConverter[] { new DictionaryConverter() });
-
-                        var envelope = (Dictionary<string, object>)doc["Envelope"];
-                        var header = (Dictionary<string, object>)envelope["WARC-Header-Metadata"];
-                        var type = (string)header["WARC-Type"];
-
-                        if (type == "response")
-                        {
-                            var payloadMetaData = (Dictionary<string, object>)envelope["Payload-Metadata"];
-                            var response = (Dictionary<string, object>)payloadMetaData["HTTP-Response-Metadata"];
-                            var url = new Uri(Uri.UnescapeDataString((string)header["WARC-Target-URI"]));
-                            string title = null;
-                            string description = null;
-
-                            if (response.ContainsKey("HTML-Metadata"))
-                            {
-                                var htmlMetaData = (Dictionary<string, object>)response["HTML-Metadata"];
-
-                                if (htmlMetaData.ContainsKey("Head"))
-                                {
-                                    var head = (Dictionary<string, object>)htmlMetaData["Head"];
-
-                                    if (head.ContainsKey("Title"))
-                                    {
-                                        title = (string)head["Title"];
-                                    }
-
-                                    if (head.ContainsKey("Metas"))
-                                    {
-                                        foreach (var meta in (IEnumerable<dynamic>)head["Metas"])
-                                        {
-                                            foreach (var prop in meta)
-                                            {
-                                                bool hasDescription = false;
-
-                                                foreach (var x in prop)
-                                                {
-                                                    if (x.Value as string == "description")
-                                                    {
-                                                        hasDescription = true;
-                                                    }
-                                                }
-
-                                                if (hasDescription && prop.Next != null)
-                                                {
-                                                    description = prop.Next.Value.ToString();
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-
-                            yield return new Dictionary<string, object>
-                                {
-                                    { "title", title },
-                                    { "description", description },
-                                    { "scheme", url.Scheme },
-                                    { "host", url.Host },
-                                    { "path", url.AbsolutePath },
-                                    { "query", url.Query },
-                                    { "url", url.ToString() },
-                                    { "filename", refFileNae}
-                                };
-                        }
-                    }
-
-                    line = reader.ReadLine();
-                }
-            }
-        }
-
         private static void Slice(string[] args)
         {
             var file = args[1];
@@ -469,7 +264,7 @@ namespace Sir.DbUtil
             //}
         }
 
-        private static void Truncate(string[] args, IStringModel model, ILoggerFactory log)
+        private static void Truncate(string[] args, IStringModel model, ILogger log)
         {
             var collection = args[1];
             var collectionId = collection.ToHash();
@@ -480,7 +275,7 @@ namespace Sir.DbUtil
             }
         }
 
-        private static void TruncateIndex(string[] args, IStringModel model, ILoggerFactory log)
+        private static void TruncateIndex(string[] args, IStringModel model, ILogger log)
         {
             var collection = args[1];
             var collectionId = collection.ToHash();
@@ -551,23 +346,6 @@ namespace Sir.DbUtil
                         yield return doc;
                         took++;
                     }
-
-                    line = reader.ReadLine();
-                }
-            }
-        }
-
-        private static IEnumerable<string> ReadAllLinesGromGz(string fileName)
-        {
-            using (var stream = File.OpenRead(fileName))
-            using (var zip = new GZipStream(stream, CompressionMode.Decompress))
-            using (var reader = new StreamReader(zip))
-            {
-                var line = reader.ReadLine();
-
-                while (!string.IsNullOrWhiteSpace(line))
-                {
-                    yield return line;
 
                     line = reader.ReadLine();
                 }
