+++ src/Sir.HttpServer/Views/Admin/Index.cshtml
@@ -1,0 +1,7 @@
+﻿@{
+    Layout = "~/Views/_Layout.cshtml";
+}
+@{ 
+    ViewBag.Title = "Admin";
+} 
+    <h2>Admin</h2>
