+++ src/Sir.HttpServer/sir.ini
@@ -3,6 +3,7 @@ cc_target_url=https://commoncrawl.org/2019/12/december-2019-crawl-archive-now-av
 cc_target_name=December 2019
 default_collection=cc_wat,cc_wet
 default_fields=title,description
+admin_password=SupervisedInformationRetrieval123!
 index_session_thread_count=32
 validate_thread_count=8
 warmup_thread_count=8
