+++ src/Sir.sln
@@ -28,7 +28,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "apps", "apps", "{B97268C5-0
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{35E2693A-1A42-4690-81A8-D424C3D24AD1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sir.Document", "Sir.Document\Sir.Document.csproj", "{F6AFA2E5-1FEA-4DC6-9386-20315126C7C4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sir.Document", "Sir.Document\Sir.Document.csproj", "{F6AFA2E5-1FEA-4DC6-9386-20315126C7C4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sir.CommonCrawl", "Sir.CommonCrawl\Sir.CommonCrawl.csproj", "{304CAC1B-825A-4D89-AE4C-9C0FC5206607}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -72,6 +74,10 @@ Global
 		{F6AFA2E5-1FEA-4DC6-9386-20315126C7C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F6AFA2E5-1FEA-4DC6-9386-20315126C7C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F6AFA2E5-1FEA-4DC6-9386-20315126C7C4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{304CAC1B-825A-4D89-AE4C-9C0FC5206607}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{304CAC1B-825A-4D89-AE4C-9C0FC5206607}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{304CAC1B-825A-4D89-AE4C-9C0FC5206607}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{304CAC1B-825A-4D89-AE4C-9C0FC5206607}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -81,6 +87,7 @@ Global
 		{610237D4-15E5-4C9F-9BAE-ACB24F634D2F} = {B97268C5-0BFA-4022-BA3F-C07C1F239C8D}
 		{D267D6FB-6DA5-444F-A628-FC0C3B1C6662} = {B97268C5-0BFA-4022-BA3F-C07C1F239C8D}
 		{03F6B627-A8CC-43A6-9BD0-5A9CE47E7EA6} = {35E2693A-1A42-4690-81A8-D424C3D24AD1}
+		{304CAC1B-825A-4D89-AE4C-9C0FC5206607} = {B97268C5-0BFA-4022-BA3F-C07C1F239C8D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {60BC16EF-AAA7-498C-A62A-56F3F740FC89}
