+++ src/Sir/ISessionFactory.cs
@@ -1,15 +1,12 @@
-﻿using Microsoft.Extensions.Logging;
-using System.IO;
+﻿using System.IO;
 
 namespace Sir
 {
     public interface ISessionFactory
     {
-        ILoggerFactory LoggerFactory { get; }
         IConfigurationProvider Config { get; }
         string Dir { get; }
         IStringModel Model { get; }
-
         Stream CreateAppendStream(string fileName, int bufferSize = 4096);
         Stream CreateAsyncAppendStream(string fileName, int bufferSize = 4096);
         Stream CreateAsyncReadStream(string fileName, int bufferSize = 4096);
